+++ network.go
@@ -38,6 +38,7 @@ var (
 type MetricCollector struct {
 	selector               *selector
 	sum                    bool
+	sumologic              bool
 	stateFile              string
 	maxRateIntervalSeconds int64
 }
@@ -45,13 +46,13 @@ type MetricCollector struct {
 // NetStats is the following: map[metric-name]map[interface-name]value
 type NetStats map[string]map[string]float64
 
-func NewCollector(includes, excludes []string, sum bool, stateFile string, maxRateIntervalSeconds int64) (*MetricCollector, error) {
+func NewCollector(includes, excludes []string, sum bool, sumologic bool, stateFile string, maxRateIntervalSeconds int64) (*MetricCollector, error) {
 	selector, err := NewDeviceSelector(includes, excludes)
 	if err != nil {
 		return nil, err
 	}
 
-	return &MetricCollector{selector, sum, stateFile, maxRateIntervalSeconds}, nil
+	return &MetricCollector{selector, sum, sumologic, stateFile, maxRateIntervalSeconds}, nil
 }
 
 func (c *MetricCollector) Collect(netStatsGetter func(*selector) (NetStats, error)) ([]*dto.MetricFamily, error) {
@@ -85,13 +86,15 @@ func (c *MetricCollector) Collect(netStatsGetter func(*selector) (NetStats, erro
 }
 
 func (c *MetricCollector) generatePromMetrics(stats NetStats, metricState *metric.CounterMetricState) []*dto.MetricFamily {
+	var sumo_family *dto.MetricFamily
 	families := make([]*dto.MetricFamily, 0)
 	nowMS := time.Now().UnixMilli()
 	metricType := "host_net"
 	help := metricHelp[metricType]
-	sumo_family := newMetricFamily(metricType, help, dto.MetricType_COUNTER)
-	families = append(families, sumo_family)
-
+	if c.sumologic {
+		sumo_family = newMetricFamily(metricType, help, dto.MetricType_COUNTER)
+		families = append(families, sumo_family)
+	}
 	for metricType, typeStats := range stats {
 		help := metricHelp[metricType]
 		if help == "" {
@@ -113,7 +116,9 @@ func (c *MetricCollector) generatePromMetrics(stats NetStats, metricState *metri
 
 		for netIF, ifValue := range typeStats {
 			counter := newCounterMetric(family, netIF, ifValue, nowMS)
-			_ = newSumoCounterMetric(sumo_family, metricType, netIF, ifValue, nowMS)
+			if c.sumologic {
+				_ = newSumoCounterMetric(sumo_family, metricType, netIF, ifValue, nowMS)
+			}
 			found, prevValue, prevTimestampMS := metricState.GetMetric(family, counter)
 			metricState.AddMetric(family, counter)
 			total += ifValue
@@ -157,7 +162,6 @@ func newCounterMetric(family *dto.MetricFamily, ifName string, value float64, ti
 	counter := &dto.Metric{
 		Label: []*dto.LabelPair{
 			{Name: &interfaceLabel, Value: &ifName},
-			//{Name: &fieldLabel, Value: &fieldName},
 		},
 		Counter: &dto.Counter{
 			Value: &value,